Filter events that rooms have declined
People can try to book a room, but if it’s already booked, then the room
will decline the attempt to book it. It still records the event in the
room’s calendar though.

Currently, the API reports double-booked meetings, which causes
confusion when 2 groups attempt to enter a room.

This change checks that a room has accepted the event before reporting
it as booked.