Make startTicker more resilient
Being unable to retrieve an OAuth 2.0 token should not be fatal error
that causes the application to exit. There are plausible reasons why
this could happen:

- intermittent DNS / network issues
- server failure

So instead, we add a recovery pattern function, and panic where
appropriate, rather than logging a fatal issue.

Also changed `startTicker` to be self-contained, and run a goroutine
which tries to `loadEvents`. This is so that only `loadEvents` needs
to have a recovery pattern.