Default to displaying events
We don’t fully understand the potential state space for events. This
change acknowledges that, and only filters events if we positively know
that we should.

Currently that just means bookings which this application can see, that
the room has rejected.

# Other states

## Private events

Events might be private, but have invited a room. This means that the
account we have can’t check to see whether the room accepted or rejected
the request. We assume that this is a valid event in that case.

## The Mike Bracken Memorial Lecture Theatre

Aka Mike’s private office.

Some room / resource calendars appear to be directly controlled by
individuals; eg 601 is not available for general booking. In that event,
we assume that the event appears in the room booking calendar, rather
than an individual’s calendar.

## User Research Lab

Some rooms / resources can be booked by people outside of GDS by another
mechanism that I’m not aware of; eg the user research lab. Again, we
assume that the booking is valid in that case.