Use proper logging library
This introduces a distinction between log levels. Currently, the logs
are very noisy and not particularly useful. Here, we introduce debug
logging for developers, versus things that operations need to know to
make the logs more actionable.


LOG_LEVEL is a new environment variable that can be used to control the
logging. By default, it’s ‘error’, but can be set to any value that
logrus supports.